server/requirements.txt
black==23.9.1
blinker==1.6.2
click==8.1.7
Flask==2.3.3
Flask-Cors==4.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
psycopg2==2.9.7
python-dotenv==1.0.0
Werkzeug==2.3.7
